requirements.txt
python-telegram-bot
python-ffmpeg